+++ lib/dirtree.ml
@@ -8,8 +8,6 @@ and node = string * t
 
 type top = Top of node
 
-let self_key = lazy (Age.Key.from_identity_file @@ Lazy.force Config.identity_file)
-
 (* always recurse on directories and keep files with extension [ext] *)
 let ext_filt ext base nm sub =
   let full = Fpath.add_seg base nm in
@@ -22,8 +20,7 @@ let ext_filt ext base nm sub =
       let name = name_of_file (Path.of_fpath full) in
       let res =
         try
-          let self_key = Lazy.force self_key in
-          match is_recipient_of_secret self_key name with
+          match is_recipient_of_secret (get_own_key ()) name with
           | false -> Skipped
           | true ->
           match decrypt_exn ~silence_stderr:true name with
@@ -35,13 +32,11 @@ let ext_filt ext base nm sub =
     else None
   with _ -> None
 
-let filt = ext_filt ext
-
 let of_path path =
   let rec sub p =
     let names = Sys.readdir (Fpath.to_string p) in
     Array.sort compare names;
-    names |> Array.to_list |> List.filter_map (fun v -> filt p v sub)
+    names |> Array.to_list |> List.filter_map (fun v -> ext_filt ext p v sub)
   in
   let sp = sub path in
   Top (Fpath.to_string path, D sp)
@@ -61,7 +56,10 @@ let to_pref_str is_last_l =
   | v :: rest -> if v then p_aux [ el ] rest else p_aux [ mid ] rest
   | [] -> ""
 
-let p d = printf "%s%s\n" (to_pref_str d)
+let p buf d s =
+  Buffer.add_string buf (to_pref_str d);
+  Buffer.add_string buf s;
+  Buffer.add_char buf '\n'
 
 (* call a different function for the last item *)
 let rec iter_but_one f lastf l =
@@ -78,20 +76,22 @@ let green = "32"
 
 let color c s = sprintf "\027[01;%sm%s\027[00m" c s
 
-let rec pp_node d n =
+let rec pp_node buf d n =
   match n with
   | nm, F r ->
     (match r with
-    | Succeeded _ -> p d (color green nm)
-    | Failed _ -> p d (color red nm)
-    | _ -> p d nm)
+    | Succeeded _ -> p buf d (color green nm)
+    | Failed _ -> p buf d (color red nm)
+    | _ -> p buf d nm)
   | nm, D nl ->
-    let () = p d (color blue nm) in
-    iter_but_one (pp_node (false :: d)) (pp_node (true :: d)) nl
+    let () = p buf d (color blue nm) in
+    iter_but_one (pp_node buf (false :: d)) (pp_node buf (true :: d)) nl
 
 let pp t =
+  let buf = Buffer.create 256 in
   match t with
   | Top (_, D l) ->
-    let () = printf ".\n" in
-    iter_but_one (pp_node [ false ]) (pp_node [ true ]) l
-  | Top (_, F _) -> ()
+    Buffer.add_string buf ".\n";
+    iter_but_one (pp_node buf [ false ]) (pp_node buf [ true ]) l;
+    Buffer.contents buf
+  | Top (_, F _) -> ""