+++ lib/template_ast.mli
@@ -1,0 +1,7 @@
+type ast_node =
+  | Iden of string
+  | Text of string
+
+type ast = ast_node list
+
+val to_string : ast_node -> string