+++ lib/template_lexer.mli
@@ -1,0 +1,2 @@
+module Encoding = Sedlexing.Utf8
+val token : Sedlexing.lexbuf -> Template_parser.token