+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Update refresh command to handle users and groups via the @<user_or_group> target syntax
 - Add --version flag to passage command
+- Refactor passage to be lib-first. Commands and utility functions are now available for usage by lib users
 
 ## 0.1.8 (2025-10-21)
 - deep refactor of the code